requirements.txt
# ONNX Runtime with TensorRT support
onnxruntime-gpu>=1.17.0
# tensorrt>=8.6.0  # Install from NVIDIA

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Export & Optimization
onnx>=1.15.0
onnxsim>=0.4.33
protobuf>=4.24.0

# YOLO Models
ultralytics>=8.1.0

# Optional: Quantization Tools
# onnxruntime-extensions>=0.9.0
# neural-compressor>=2.4.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter for examples
jupyter>=1.0.0
ipywidgets>=8.1.0
